fix testcase run twice times
