+++ packages/node/tests/cli.test.ts
@@ -5,7 +5,7 @@ import { fileURLToPath } from "node:url";
 import { beforeAll, describe, expect, test } from "vitest";
 
 import { GenericRespone, SubscribeDRPRequest } from "../src/proto/drp/node/v1/rpc_pb.js";
-import { run } from "../src/runner.js";
+import * as run from "../src/run.js";
 
 const protoPath = path.resolve(
 	dirname(fileURLToPath(import.meta.url)),
@@ -21,7 +21,7 @@ describe("Run DRP with cli", () => {
 	let client: any;
 
 	beforeAll(async () => {
-		await run();
+		await run;
 		client = new service.DrpRpcService(`localhost:6969`, grpc.credentials.createInsecure());
 	});
